+++ tests/unit/test_docset.py
@@ -7,13 +7,24 @@ sys.path.append(os.path.relpath(__file__) + "/../src")
 import src.zeal  # NOQA: E402
 
 
-def test_list_docsets():
+def test_download_docset():
+    with tempfile.TemporaryDirectory() as data_dir:
+        feeds_dir = os.path.join(data_dir, "feeds")
+        os.mkdir(feeds_dir)
+        docset_dir = os.path.join(data_dir, "docsets")
+        os.mkdir(docset_dir)
+        feeds_path = src.zeal.downloads.get_feeds(feeds_dir)
+        src.zeal.docset.download("Django", feeds_path, docset_dir=docset_dir)
+        assert os.path.isdir(os.path.join(docset_dir, "Django.docset"))
+
+
+def test_list_all():
     with tempfile.TemporaryDirectory() as docset_dir:
         os.mkdir(os.path.join(docset_dir, "docset1.docset"))
         os.mkdir(os.path.join(docset_dir, "docset2.docset"))
         os.mkdir(os.path.join(docset_dir, "docset3.docset"))
         os.mkdir(os.path.join(docset_dir, "not_a_docset"))
-        docset_list = src.zeal.list_docsets.get_list(docset_dir=docset_dir)
+        docset_list = src.zeal.docset.list_all(docset_dir=docset_dir)
         assert "docset1" in docset_list
         assert "docset2" in docset_list
         assert "docset3" in docset_list
